requirements.txt
# requirements.txt
yfinance
pandas
numpy
pandas_ta
beautifulsoup4
requests
lxml
matplotlib
mplfinance
vaderSentiment
groq
python-dotenv
